requirements.txt
PyJWT
requests
pytest
psycopg2-binary
celery
redis
pydantic-settings
